+++ apps/app/src/app/(app)/[orgId]/risk/(overview)/page.tsx
@@ -108,7 +108,7 @@ const getAssignees = cache(async () => {
       organizationId: session.session.activeOrganizationId,
       isActive: true,
       role: {
-        notIn: ['employee'],
+        notIn: ['employee', 'contractor'],
       },
     },
     include: {