+++ packages/db/prisma/migrations/20251112145246_add_deactivated_to_member/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Member" ADD COLUMN     "deactivated" BOOLEAN NOT NULL DEFAULT false;